requirements.txt
PyYAML
httpx